pre-commit: add explicit dependencies to (node) hooks
and a script to sync from package-lock to the pre-commit config.
This allows us to enable pre-commit.ci, which seems to be very fast and
also provides automatic fixes to PRs.

Currently, this does not seem to work for eslint, which always fails for
svelte components (Typescript does not seem to be compiled).